feat: refactor plugin to use main repo's plugin structure
BREAKING CHANGE:
Use standard plugins instead of proposed `getCommits` plugin.

The previous plugin implementation was based around a [proposed, but
rejected `getCommits`
plugin](https://github.com/semantic-release/semantic-release/pull/531).

The new implementation is functionally equivalent, but is slower as it
has to repeat the filtering of commits an additional time. Rather than
doing it once, as in the proposed `getCommits`, we need to do it for
both `analyzeCommits` and `generateNotes` separately. Could probably be
optimized through memoization or global variable.